backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
groq==0.4.2
requests==2.31.0
python-dotenv==1.0.0
feedparser==6.0.10
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
beautifulsoup4==4.12.0